refactor: update imports to use shadcn directory in indents pages
